+++ Assets/Scripts/SGG.PerfMeter/Runtime/AssemblyInfo.cs
@@ -2,7 +2,7 @@ using System.Reflection;
 using System.Runtime.CompilerServices;
 
 [assembly: AssemblyMetadata("SGG.PerfMeter.PackageName", "com.sungeargames.perfmeter")]
-[assembly: AssemblyMetadata("SGG.PerfMeter.PackageVersion", "2026.7.16-1")]
+[assembly: AssemblyMetadata("SGG.PerfMeter.PackageVersion", "2026.7.19-1")]
 [assembly: InternalsVisibleTo("SGG.PerfMeter.Tests.EditMode")]
 [assembly: InternalsVisibleTo("SGG.PerfMeter.Editor")]
 [assembly: InternalsVisibleTo("SGG.PerfMeter.URP")]